+++ packages/react/src/hooks/index.ts
@@ -6,7 +6,13 @@ export type {
   PlColorSchemeResult,
   PlResolvedColorScheme
 } from './usePlColorScheme.js';
+export { usePlDisclosure } from './usePlDisclosure.js';
+export type { PlDisclosureResult } from './usePlDisclosure.js';
+export { usePlElementSize } from './usePlElementSize.js';
+export type { PlElementSize } from './usePlElementSize.js';
 export { usePlHotKeys } from './usePlHotKeys.js';
 export type { PlHotKeysOptions, PlHotKeysTarget } from './usePlHotKeys.js';
 export { usePlMediaQuery } from './usePlMediaQuery.js';
+export { usePlOnScreen } from './usePlOnScreen.js';
+export type { PlOnScreenOptions } from './usePlOnScreen.js';
 export { usePlReducedMotion } from './usePlReducedMotion.js';